requirements.txt
amqp==2.5.0
asgiref==3.3.1
asn1crypto==1.4.0
Beaker==1.11.0
beautifulsoup4==4.9.1
billiard==3.6.0.0
boto3==1.9.209
botocore==1.12.209
cachetools==4.2.1
celery==4.3.0
certifi==2019.6.16
cffi==1.14.5
chardet==3.0.4
cliff==3.1.0
cmd2==0.8.9
contextlib2==0.5.5
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
cryptography==2.6.1
defusedxml==0.6.0
Django==2.2.2
django-admin-daterange-filter==0.1.3
django-admin-rangefilter==0.5.2
django-cors-headers==3.6.0
django-admin-rangefilter==0.5.2
django-debug-toolbar
django-extensions==2.1.9
django-filter==2.1.0
django-redis==4.10.0
django-rest-swagger==2.2.0
django-storages==1.7.1
django-templated-mail==1.1.1
djangorestframework==3.9.4
djangorestframework-simplejwt==4.6.0
docutils==0.15.2
drf-yasg==1.17.1
http-ece==1.1.0
httplib2==0.19.0
idna==2.8
inflection==0.5.1
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
kombu==4.6.3
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
pbr==5.4.5
prettytable==0.7.2
psycopg2==2.8.3
psycopg2-binary==2.8.3
py-vapid==1.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pyperclip==1.8.0
python-dateutil==2.8.0
python-http-client==3.3.2
python3-openid==3.2.0
pytz==2019.1
pywebpush==1.9.4
PyYAML==5.4.1
redis==3.5.3
reportlab==3.5.59
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.7.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
s3transfer==0.2.1
schema==0.7.4
sendgrid==6.1.0
sentry-sdk==0.20.3
simplejson==3.17.2
six==1.12.0
soupsieve==2.0.1
sqlparse==0.3.0
stevedore==3.3.0
typing-extensions==3.7.4.1
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
waitress==1.4.4
wcwidth==0.2.3
WebOb==1.8.6
